+++ src/Components/BookSearchForm/index.css
@@ -1,0 +1,3 @@
+.search-book-result {
+    margin: auto;
+}
