+++ src/index.tsx
@@ -3,7 +3,6 @@ import * as ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import BookSearchForm from './components/page/Header';
 
 ReactDOM.render(
   <React.StrictMode>
